services/submission-api/requirements.txt
fastapi==0.115.*
psycopg2-binary==2.9.*
uvicorn[standard]==0.30.*
minio==7.*
python-dotenv==1.0.*
